requirements.txt
APScheduler==3.6.3
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
comtypes==1.1.10
distlib==0.3.2
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.0.12
html5lib==1.1
idna==2.10
IMAPClient==2.2.0
importlib-metadata==4.6.1
iso3166==2.0.2
Js2Py==0.71
lxml==4.6.3
MouseInfo==0.1.3
numpy==1.19.1
openpyxl==3.0.7
packaging==20.9
Pillow==8.3.1
pipwin==0.5.1
platformdirs==2.0.2
psutil==5.8.0
PyAutoGUI==0.9.52
pycodestyle==2.8.0
PyGetWindow==0.0.9
pyjokes==0.6.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.27
pySmartDL==1.3.4
python-docx==0.8.11
python-telegram-bot==13.10
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywin32==301
pyzmail==1.0.3
pyzmail36==1.0.4
requests==2.25.1
scipy==1.7.0
selenium==3.141.0
six==1.16.0
sort-dataframeby-monthorweek==0.4
sorted-months-weekdays==0.2
soupsieve==2.2.1
SpeechRecognition==3.8.1
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
virtualenv==20.6.0
webencodings==0.5.1
wikipedia==1.4.0
zipp==3.5.0